Reference

Contents

Reference

Contents

Index

Index

BlockmodelUtils.Blockmodel

Base.map

BlockmodelUtils.blockmodel

BlockmodelUtils.densityplot

BlockmodelUtils.densityplot!

BlockmodelUtils.flowerplot

BlockmodelUtils.flowerplot!

BlockmodelUtils.permuteplot

BlockmodelUtils.permuteplot!

Graphs.density

BlockmodelUtils.Blockmodel — Type

Struct representing a graph's blocked adjacency matrix.

Base.map — Method

map(fun, bm::Blockmodel)

Map function fun over the blocks of the Blockmodel bm and return a k by k matrix, where k is the number of groups constituting the blockmodel.

Example

using Graphs, BlockmodelUtils

g = erdos_renyi(20, 0.1)
gs = rand(['a', 'b'], 20)
bm = blockmodel(g, gs)

map(sum, bm) # number of edges in each block

BlockmodelUtils.blockmodel — Method

blockmodel(g, groups; by=identity)

Create a Blockmodel from Graphs.AbstractGraph g and vector of group memberships groups. The function passed to by used to order the groups in the blockmodel.

Example

```julia using Graphs, BlockmodelUtils

g = erdos_renyi(20, 0.1) gs = rand(['a', 'b'], 20) bm = blockmodel(g, gs)

BlockmodelUtils.densityplot — Function

densityplot!(ax, bm)
densityplot(bm)

Plot the block density matrix as a heatmap.

Attributes

xticklabels = bm.labels

yticklabels = bm.labels

rotate_xlabels = false

kwargs... are passed to Makie.heatmap

BlockmodelUtils.densityplot! — Function

densityplot!(ax, bm)
densityplot(bm)

Plot the block density matrix as a heatmap.

Attributes

xticklabels = bm.labels

yticklabels = bm.labels

rotate_xlabels = false

kwargs... are passed to Makie.heatmap

BlockmodelUtils.flowerplot — Function

flowerplot!(ax, bm)
flowerplot(bm)

Plot the network with nodes positioned in circles according to the blockmodel groups. The first group is placed in the center.

Attributes

edgecolor = :grey70

edgewidth = 1

nodecolor = :black

showlabels = true

radii = fill(1, length(bm.labels))

args... further arguments are passed to Makie.scatter

BlockmodelUtils.flowerplot! — Function

flowerplot!(ax, bm)
flowerplot(bm)

Plot the network with nodes positioned in circles according to the blockmodel groups. The first group is placed in the center.

Attributes

edgecolor = :grey70

edgewidth = 1

nodecolor = :black

showlabels = true

radii = fill(1, length(bm.labels))

args... further arguments are passed to Makie.scatter

BlockmodelUtils.permuteplot — Function

permuteplot!(ax, bm)
permuteplot(bm)

Plot the permuted adjacency matrix with a grid outlining the blockmodel blocks.

Attributes

cellcolor = :black

linecolor = :black

framecolor = linecolor

rotate_xlabels = fale

xticklabels = bm.labels

yticklabels = bm.labels

BlockmodelUtils.permuteplot! — Function

permuteplot!(ax, bm)
permuteplot(bm)

Plot the permuted adjacency matrix with a grid outlining the blockmodel blocks.

Attributes

cellcolor = :black

linecolor = :black

framecolor = linecolor

rotate_xlabels = fale

xticklabels = bm.labels

yticklabels = bm.labels

Graphs.density — Method

density(bm::Blockmodel)

Compute the blockdensity matrix for Blockmodel bm. For diagonal blocks, the diagonal is subtracted from the number of possible edges.